♻️ refactor: 폴더 구조 변경 (Jira Ticket : CCC-78)
