+++ app/(board)/board/[id]/components/BoardDetailDescription.tsx
