+++ app/(board)/boards/components/SearchBoard.tsx
